Started pathplanner autos
None of the path final destinations are tuned(or correct). This for now only serves to construct a shell or scaffold to build off of. For the command: score coral_ex, it might be prudent to couple that with an "a to ab" path for example(with a parallel race group).