SysID subsystems (#73)
* jank implementation of sysID

Not tested implementation, should logically work. It does run.

* Elevator implementation

I doubt that this will work because it doesn't account for gravity, but the rough implementation is added. It builds .

* arm Implementation

I'd bet my left kidney that this doesn't work because of gravity, but it at least gets the framework set up.

* slight simulator testing

Elevator sysID 100% does not work for either quasistatic or dynamic, arm sysID seems normal enough(but it needs a bound maybe) for quasistatic, and the dynamic moves really fast. The swerve sysID seems like it works. All of these were tested in simulator.

* Fixed elevator

The elevator not working is fixed. Same problems as with arm

* Added safety zones for the coral subSystem

safety zones for SysID

* Removed dynamic

Dynamic was pumping 7 volts in, thus it was removed. Theoretically should work on real robot.

* Preparing to test open loop

The hardware has changed so we're just preparing to run open loop before the sys ID

* slight tweaks to arm speed for open loop

* updated arm offset

Arm encoder reports incorrect values. The base position is correct because of offset, but it overshoots at max height. I'm moving on to elevator.

* Elevator works but height is slightly different

Elevator works but the height is slightly higher than we expected.

* Manually offset arm

Arm is now reporting correct values(although the starting position is slightly high)

* Tested closed loop

It worked, however, sys ID did not for the arm and elevator. It didn't run.

* SysID didn't work

Manually tuned PID, every subsystem roughly works

* "fixed" code

not sure if this worse.

* More manual PID tuning

Code works(pretty well) as is, although the intake command doesn't run. Tweaking values so that that happens next, and if that's accomplished, then we can make it faster(ideally in a second or two)

* changes for PID tuning

Also raised home position to create more thrusting force whilst intaking.

* Minor manual tuning

Not the correct values, just slight tweaks for more better poi code. The elevator still doesn't approach the correct position fast enough

* precaution

* temporarily changes

For the sake of the sys-id, open loop will temporarily become half closed loop. We'll see if this runs

* It kinda worked

My first test, quasistatic worked fine, however, I ran dynamic and the robot started turning, and now, whenever I feed it a horizontal movement, two wheels start spinning endlessly.

* Cleaned up code

Cleaned up code and fudge values.

* Making things sane

Changed endState to setPointState,
changed targetState to setPointState

* Updated Elastic layout

Updated elastic layout to have the
coral subsystem commands

* Debugging jump in arm and elevator

The target state is jumping in arm and elevator. I believe it's a trapezoid profile problem because the values we're feeding in are normal. I'm going to try turning down the constraints

* More problems

Likely a multitude of problems. Sometimes the setPoint randomly jumps up for some reason. It doesn't seem like the elevator position is incorrect however.

* fudge tuning

Not good tuning, also tuned for simulator. Please fix over spring break

* Our readings were "wrong"

Finding new values and slightly adjusting them. Also, since we swapped encoders, there's a change.

* removed smart dashboard log

removed a log