reject measurements from a closer distance
this will help prevent the lime lights on the front of the robot from picking up the reef all the way at the intake station. also reduced the max rotational speed of the auto align to match path planner rotational config. Also adjusted the navgrid to have more padding to avoid going too close to the reef